+++ client/packages/cli/src/commands/submit/sfx.ts
@@ -14,7 +14,7 @@ import { EncodedArgs, SideEffect, TransferEncodedArgs } from "@/schemas/sfx.ts"
 
 export const spinner = ora()
 
-export const handleSubmitSfxCmd = (sfxFile: string) => {
+export const handleSubmitSfxCmd = (sfxFile: string, exportMode: boolean) => {
   const unvalidatedExtrinsic = readSfxFile(sfxFile)
 
   if (!unvalidatedExtrinsic) {
@@ -29,7 +29,7 @@ export const handleSubmitSfxCmd = (sfxFile: string) => {
     process.exit(1)
   }
 
-  submitSfx(extrinsic)
+  submitSfx(extrinsic, exportMode)
 }
 
 export const readSfxFile = (filePath: string) => {
@@ -46,7 +46,7 @@ export const readSfxFile = (filePath: string) => {
   }
 }
 
-export const submitSfx = async (extrinsic: Extrinsic) => {
+export const submitSfx = async (extrinsic: Extrinsic, exportMode: boolean) => {
   const config = getConfig()
 
   spinner.text = "Submitting extrinsic..."
@@ -56,7 +56,7 @@ export const submitSfx = async (extrinsic: Extrinsic) => {
     process.exit(1)
   }
 
-  const { circuit, sdk } = await createCircuitContext()
+  const { circuit, sdk } = await createCircuitContext(exportMode)
   const transactionArgs = buildSfx(
     circuit,
     extrinsic.sideEffects,
@@ -66,7 +66,9 @@ export const submitSfx = async (extrinsic: Extrinsic) => {
 
   try {
     const transaction = circuit.tx.circuit.onExtrinsicTrigger(
-      transactionArgs.sideEffects,
+      transactionArgs.sideEffects as Parameters<
+        typeof circuit.tx.circuit.onExtrinsicTrigger
+      >[0],
       transactionArgs.speed_mode
     )
     const submissionHeight = await sdk.circuit.tx.signAndSendSafe(transaction)