+++ node/standalone/src/chain_spec.rs
@@ -176,9 +176,14 @@ fn testnet_genesis(
                 .collect(),
         },
         treasury: Default::default(),
+        escrow_treasury: Default::default(),
+        fee_treasury: Default::default(),
+        parachain_treasury: Default::default(),
+        slash_treasury: Default::default(),
         aura: AuraConfig {
             authorities: initial_authorities.iter().map(|x| (x.0.clone())).collect(),
         },
+        rewards: Default::default(),
         grandpa: GrandpaConfig {
             authorities: initial_authorities
                 .iter()
@@ -197,6 +202,7 @@ fn testnet_genesis(
         },
         contracts_registry: Default::default(),
         account_manager: Default::default(),
+        attesters: Default::default(),
         clock: Default::default(),
         three_vm: Default::default(), // TODO: genesis for this needs to be setup for the function pointers\
         evm: EvmConfig {