+++ primitives/src/lib.rs
@@ -41,6 +41,7 @@ use sp_std::{prelude::*, vec};
 use std::fmt::Debug;
 
 pub mod account_manager;
+pub mod attesters;
 pub mod circuit;
 pub mod claimable;
 pub mod clock;
@@ -87,6 +88,22 @@ impl GatewayType {
             Self::TxOnly(nonce) => nonce,
         }
     }
+}
+
+#[derive(Clone, Eq, PartialEq, Encode, Decode, Debug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Default)]
+pub enum TreasuryAccount {
+    #[default]
+    Treasury,
+    Escrow,
+    Fee,
+    Parachain,
+    Slash,
+}
+
+pub trait TreasuryAccountProvider<Account> {
+    fn get_treasury_account(treasury_account: TreasuryAccount) -> Account;
 }
 
 #[derive(Clone, Eq, PartialEq, Encode, Decode, Debug, TypeInfo)]